#39 fix tailwind適用
